move cli to correct folder
