+++ cmd/cli/commands_test.go
